modified pages/Elements/ButtonOpenAnAccountMainBanner
